+++ pyturbo/systems/channel/__init__.py
@@ -1,0 +1,6 @@
+from pyturbo.systems.channel.channel_aero import ChannelAero
+from pyturbo.systems.channel.channel_geom import ChannelGeom
+
+from pyturbo.systems.channel.channel import Channel  # isort: skip
+
+__all__ = ["ChannelAero", "ChannelGeom", "Channel"]