+++ pyturbo/systems/structures/__init__.py
@@ -1,8 +1,4 @@
-from pyturbo.systems.structures.channel import Channel
-from pyturbo.systems.structures.channel_aero import ChannelAero
-from pyturbo.systems.structures.channel_geom import ChannelGeom
 from pyturbo.systems.structures.core_cowl import CoreCowl
-from pyturbo.systems.structures.fan_duct import FanDuct
 from pyturbo.systems.structures.intermediate_casing import IntermediateCasing
 
-__all__ = ["ChannelAero", "ChannelGeom", "Channel", "IntermediateCasing", "FanDuct", "CoreCowl"]
+__all__ = ["IntermediateCasing", "CoreCowl"]